sorted components in folders
