+++ src/components/admin/Tables/PendingTable.js
@@ -2,40 +2,40 @@
 import * as React from 'react';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
-import Toolbar from '@material-ui/core/Toolbar';
 import TableCell from '@material-ui/core/TableCell';
 import TableContainer from '@material-ui/core/TableContainer';
+import Toolbar from '@material-ui/core/Toolbar';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
+import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
 import TableFooter from '@material-ui/core/TableFooter';
 import TablePagination from '@material-ui/core/TablePagination';
+import styles from '../admin.module.css';
+import {Typography} from '@material-ui/core';
 import IconButton from '@material-ui/core/IconButton';
 
-import styles from './admin.module.css';
 import {TablePaginationActions,
   computeEmptyRows,
-  computeRows} from './TablePaginationActions';
+  computeRows} from '../TablePaginationActions';
 
-import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
-import {Typography} from '@material-ui/core';
+import PendingInfo from '../Dialogs/PendingInfo';
 
-import ActionInfo from './ActionInfo';
+function createData(id: number, name: string,
+    email: string, date: any): {id: number, name: string,
+    email: string, date: any} {
+  return {id, name, email, date};
+}
 
-const rows: Array<any> = Array.from([
-  '[NOOGLER CHECK 1]: A partner of a noogler requires access',
-  '[DATABASE]: A new member could not be added to the database.',
-  '[NOOGLER CHECK 2]: A partner of a noogler requires access',
-], (message: string, id: number): {id: number, message: string, date: any} => ({
-  id,
-  message,
-  date: new Date(),
-}));
+const rows = [
+  createData(1, 'Alice Joy', 'alicee@gmail.com', new Date()),
+  createData(2, 'David Toms', 'dt@yahoo.com', new Date()),
+];
 
-export default function ActionTable(): React.Node {
+export default function PendingTable(): React.Node {
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
-  const [selectedRow, setSelectedRow] = React.useState(-1);
+  const [selectedPending, setSelectedPending] = React.useState(-1);
 
   const handleChangePage = (event: any, newPage: number) => {
     setPage(newPage);
@@ -46,50 +46,47 @@ export default function ActionTable(): React.Node {
     setPage(0);
   };
 
-  const handleSelectedRow = (event: any, rowId: number) => {
-    console.log(rowId);
-    setSelectedRow(rowId);
+  const handleOpenDialog = (event: any, rowId: number) => {
+    setSelectedPending(rowId);
   };
 
-  const handleCloseModal = () => {
-    setSelectedRow(-1);
+  const handleCloseDialog = (event: any) => {
+    setSelectedPending(-1);
   };
 
   return (
     <Paper>
-      <Toolbar className={styles.titleActionBar} variant='dense'>
-        <Typography className={styles.titleText} variant='h6'>
-          Immediate action required
+      <Toolbar className={styles.titlePendingBar} variant='dense'>
+        <Typography variant='h6'>
+          Pending requests
         </Typography>
       </Toolbar>
       <TableContainer>
-        <Table aria-label='Immediate actions' size='small'>
+        <Table aria-label='Pending memberships' size='small'>
           <TableHead>
             <TableRow>
-              <TableCell>Date</TableCell>
-              <TableCell>Message</TableCell>
+              <TableCell>Name</TableCell>
+              <TableCell>Email</TableCell>
               <TableCell></TableCell>
             </TableRow>
           </TableHead>
-
           <TableBody>
             {computeRows(page, rows, rowsPerPage)
                 .map((row: any): React.Node => (
                   <TableRow key={row.id}>
-                    <TableCell>{row.date.toDateString()}</TableCell>
-                    <TableCell>{row.message}</TableCell>
+                    <TableCell>{row.name}</TableCell>
+                    <TableCell>{row.email}</TableCell>
                     <TableCell>
                       <IconButton
                         onClick={(event: any): void =>
-                          handleSelectedRow(event, row.id)}>
-                        <MoreHorizIcon />
+                          handleOpenDialog(event, row.id)}>
+                        <MoreHorizIcon/>
                       </IconButton>
                     </TableCell>
-                    <ActionInfo
-                      action={row}
-                      open={selectedRow === row.id}
-                      onClose={handleCloseModal} >
-                    </ActionInfo>
+                    <PendingInfo
+                      user={row}
+                      open={row.id === selectedPending}
+                      onClose={handleCloseDialog} />
                   </TableRow>
                 ))}
